First "stable" build Added "lib" crate type back to be able to use functions from the lib within the binary again (didn't realize I could just do that before).
Added #[no_mangle] so the functions exported properly.

Pushed the `eventlog::init` call into a thread to delay it. Without this, RPC is not available as LSASS is loading the DLL very early in the boot process.

Changed `replace_symbols` to operate in-place as a minor optimization.

Changed out from using the `winreg` crate to the official `windows_registry` crate.

Changed the known-good password test to use a randomly generated password.